fix: change header to client and remove weird layout directory
