--- src/app/(header)/layout.tsx
@@ -1,10 +1,0 @@
-import Header from "@/header";
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return (
-    <>
-      <Header />
-      {children}
-    </>
-  );
-}
