+++ src/app/bus/[busId]/page.tsx
