+++ src/app/layout.tsx
@@ -4,6 +4,7 @@ import { Inter } from "next/font/google";
 import { cookies } from "next/headers";
 
 import Footer from "@/footer";
+import Header from "@/header";
 import { PostHogProvider } from "@/posthog";
 import { ServiceInfoProvider } from "@/serviceinfo";
 import { ClerkProvider } from "@clerk/nextjs";
@@ -11,6 +12,7 @@ import { dark } from "@clerk/themes";
 import { type Metadata } from "next";
 import { TRPCReactProvider } from "t/react";
 import { Toaster } from "~/components/ui/sonner";
+import { AlertNavBtn } from "./_components/alertNavigationBtn";
 
 const inter = Inter({
   subsets: ["latin"],
@@ -44,6 +46,7 @@ export default async function RootLayout({
               className={cookie.get("theme")?.value === "dark" ? "dark" : ""}
             >
               <body className={`font-sans ${inter.variable}`}>
+                <Header serviceNavigation={<AlertNavBtn />} />
                 {children}
                 <Footer />
                 <Toaster />
